+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - [fix:fix no registry when lossless is disabled.](https://github.com/Tencent/spring-cloud-tencent/pull/1311)
 - [fix:fix the ratelimit bug for 2022](https://github.com/Tencent/spring-cloud-tencent/pull/1314)
 - [feat:add Tencent Cloud TSF support.](https://github.com/Tencent/spring-cloud-tencent/pull/1317)
+- [feat:support consul config.](https://github.com/Tencent/spring-cloud-tencent/pull/1319)